Proposed changed to allow multiple linked days.
This is a tentative change since the implications of this change
affect a wide-array of things in the application.

Reports rely on Linked Days, Various displays rely on it, calculations
etc. The proposed change has included a boatload of tests to allow try
to ensure that we don't step on any of the meaninful toes, the rest
can be discovered as and when. This sounds cavalier but I am not a
psychic.